+++ packages/console/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.22.0](https://github.com/deephaven/web-client-ui/compare/v1.21.1...v1.22.0) (2026-06-10)
+
+### Bug Fixes
+
+- DH-22318: incorrect typing in ConsoleContainer props and ConsolePanel handler signatures ([#2680](https://github.com/deephaven/web-client-ui/issues/2680)) ([cf1f1c8](https://github.com/deephaven/web-client-ui/commit/cf1f1c89c3abc3e8587c5ea6ec62f30fe1e3aff7))
+
 ## [1.21.1](https://github.com/deephaven/web-client-ui/compare/v1.21.0...v1.21.1) (2026-06-04)
 
 **Note:** Version bump only for package @deephaven/console